backend/requirements.txt

boto3==1.35.0
uvicorn==0.23.2
fastapi==0.101.1
python-multipart
python-dotenv==1.0.1
requests==2.32.3
python-swiftclient==4.1.0
PyJWT==2.7.0
python-jose[cryptography]==3.3.0
sqlalchemy==2.0.22
passlib[bcrypt]
argon2-cffi==21.3.0